Create CommandlineDebugging
As GDBstub library from version 2.0 on,  as far as I know , is not capable to execute source level debugging. I have included an old version of the gdbstub library in my projects. The library is based on the toolchain of VisualGdb. This document decribes  how to use gdbstub for beginners.